backend/requirements.txt
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
greenlet==3.3.2
groq==1.1.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.7
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycparser==3.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.19.2
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
rich==14.3.3
rich-toolkit==0.19.7
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
typer==0.24.1
typing_extensions==4.15.0
ujson==5.11.0
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
